requirements.txt
# Core training stack
unsloth
transformers>=4.41.0
datasets
trl
accelerate
peft

torch
pandas

# Common HF/tokenizer runtime deps
sentencepiece
protobuf
safetensors
huggingface_hub
requests
tqdm
python-dotenv
telethon
